feat: only expose Highlight component
