+++ packages/highlighting-snippeting-vdom/src/index.ts
@@ -1,2 +1,1 @@
 export * from './Highlight';
-export * from './Snippet';
